Hud is values are updated

+++ include/utils/spring.h
@@ -2,7 +2,7 @@
 
 typedef struct Spring
 {
-    int restLength;
+    double restLength;
     double damping;
     double velocity;
     double force;
